MACsec: If MACsec block is not powered on, returning appropriate error code
Due to race conditions, it might happen that the application thinks that the
MACsec block in the PHY is powered on, whereas it actually is powered off.
When it's powered off, the old implementation threw a trace error and returned
VTSS_RC_ERROR.
This has now been changed to throwing a trace info while returning
VTSS_RC_ERR_MACSEC_PHY_POWERED_DOWN. The application can then use this error
code to take proper action.
This is only implemented for the old VTSS PHYs and not the LAN80xx ones.