MACsec: The egress property always_include_sci was used in ingress as well
If the SecY's configuration or the Tx SC's configuration had always_include_sci
set, the previous code also required incoming frames to include an SCI in the
SecTAG. This is not according to the IEEE802.1AE standard, which states nothing
about the use of always_include_sci ingress-wise.
The new code does not require a match on SCI ingress-wise if always_include_sci
is set in the configuration. Instead, the user may require a match by calling
mepa_macsec_pattern_set() with the match flags including a new match flag
called MEPA_MACSEC_MATCH_SCI. This flag is only valid if the direction is set to
MEPA_MACSEC_DIRECTION_INGRESS and the action is set to
MEPA_MACSEC_MATCH_ACTION_CONTROLLED_PORT.